Adds window.width check
This commit conditionally renders the graph based on the initial width
of the window.  If the window is smaller than 992px it renders with
sticky divs, else it renders normally.

Its been a slog to work this out; the syntax for react-sticky was
particularly fragile. For example this breaks with out much explanation:
```
<Sticky> {({styles}) => (...)}
```
But this works:
```
<Sticky>
    {({styles}) => (...)}
```

That lesson took me a few hours to sort out.

Then theres the responsive issue. The sticky div doesn't work on smaller
viewports, but there's no way to conditionally apply it.  I've set up
one graph here to review, but its basicaly a big honking 200 line if statement.

In the course of this looked at two other approaches. First, I tried a
jQuery based library called Scroll Magic, but that broke becuase of how
React interacts with the DOM. I also tried a pure css approach using
Bootstrap helper classes, and inlining `position: sticky`. Neither
worked.